+++ src/main/java/com/whatpl/domain/chat/repository/message/ChatMessageQueryRepository.java
@@ -1,4 +1,4 @@
-package com.whatpl.domain.chat.repository;
+package com.whatpl.domain.chat.repository.message;
 
 import com.whatpl.domain.chat.dto.ChatMessageDto;
 import org.springframework.data.domain.Pageable;
