+++ src/main/java/com/whatpl/domain/chat/repository/room/ChatRoomRepository.java
@@ -1,4 +1,4 @@
-package com.whatpl.domain.chat.repository;
+package com.whatpl.domain.chat.repository.room;
 
 import com.whatpl.domain.chat.domain.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
