+++ src/main/java/com/whatpl/domain/chat/service/ChatMessageService.java
@@ -3,8 +3,8 @@ package com.whatpl.domain.chat.service;
 import com.whatpl.domain.chat.domain.ChatMessage;
 import com.whatpl.domain.chat.domain.ChatRoom;
 import com.whatpl.domain.chat.dto.ChatMessageDto;
-import com.whatpl.domain.chat.repository.ChatMessageRepository;
-import com.whatpl.domain.chat.repository.ChatRoomRepository;
+import com.whatpl.domain.chat.repository.message.ChatMessageRepository;
+import com.whatpl.domain.chat.repository.room.ChatRoomRepository;
 import com.whatpl.global.exception.BizException;
 import com.whatpl.global.exception.ErrorCode;
 import com.whatpl.domain.member.domain.Member;
