+++ src/main/java/com/whatpl/domain/chat/service/ChatRoomService.java
@@ -2,7 +2,7 @@ package com.whatpl.domain.chat.service;
 
 import com.whatpl.domain.chat.domain.ChatRoom;
 import com.whatpl.domain.chat.dto.ChatRoomResponse;
-import com.whatpl.domain.chat.repository.ChatRoomRepository;
+import com.whatpl.domain.chat.repository.room.ChatRoomRepository;
 import com.whatpl.global.util.PaginationUtils;
 import com.whatpl.domain.project.domain.Apply;
 import com.whatpl.domain.project.model.ApplyType;
