+++ src/main/java/com/whatpl/domain/project/repository/apply/ApplyRepository.java
@@ -1,8 +1,7 @@
-package com.whatpl.domain.project.repository;
+package com.whatpl.domain.project.repository.apply;
 
 import com.whatpl.domain.member.domain.Member;
 import com.whatpl.domain.project.domain.Apply;
-import com.whatpl.domain.project.domain.Project;
 import com.whatpl.global.common.model.ApplyStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,8 +9,6 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.Optional;
 
 public interface ApplyRepository extends JpaRepository<Apply, Long> {
-
-    Optional<Apply> findByProjectAndApplicant(Project project, Member applicant);
 
     @Query("select a from Apply a left join fetch a.project left join fetch a.applicant where a.id = :id")
     Optional<Apply> findWithProjectAndApplicantById(Long id);
