+++ src/main/java/com/whatpl/domain/project/repository/participant/ProjectParticipantRepository.java
@@ -1,6 +1,5 @@
-package com.whatpl.domain.project.repository;
+package com.whatpl.domain.project.repository.participant;
 
-import com.whatpl.global.common.model.Job;
 import com.whatpl.domain.project.domain.ProjectParticipant;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,8 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface ProjectParticipantRepository extends JpaRepository<ProjectParticipant, Long> {
-
-    int countByProjectIdAndJob(Long projectId, Job job);
 
     @EntityGraph(attributePaths = {"project", "participant"})
     Optional<ProjectParticipant> findWithAllById(Long participantId);
