+++ src/main/java/com/whatpl/domain/project/service/ProjectApplyService.java
@@ -11,8 +11,8 @@ import com.whatpl.domain.project.domain.Apply;
 import com.whatpl.domain.project.domain.Project;
 import com.whatpl.domain.project.dto.ApplyResponse;
 import com.whatpl.domain.project.dto.ProjectApplyRequest;
-import com.whatpl.domain.project.repository.ApplyRepository;
-import com.whatpl.domain.project.repository.ProjectRepository;
+import com.whatpl.domain.project.repository.apply.ApplyRepository;
+import com.whatpl.domain.project.repository.project.ProjectRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
