+++ src/main/java/com/whatpl/domain/project/service/ProjectCommentService.java
@@ -11,8 +11,8 @@ import com.whatpl.domain.project.dto.ProjectCommentCreateRequest;
 import com.whatpl.domain.project.dto.ProjectCommentDto;
 import com.whatpl.domain.project.dto.ProjectCommentListResponse;
 import com.whatpl.domain.project.dto.ProjectCommentUpdateRequest;
-import com.whatpl.domain.project.repository.ProjectCommentRepository;
-import com.whatpl.domain.project.repository.ProjectRepository;
+import com.whatpl.domain.project.repository.comment.ProjectCommentRepository;
+import com.whatpl.domain.project.repository.project.ProjectRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
