+++ src/main/java/com/whatpl/domain/project/service/ProjectLikeService.java
@@ -6,8 +6,8 @@ import com.whatpl.domain.member.domain.Member;
 import com.whatpl.domain.member.repository.MemberRepository;
 import com.whatpl.domain.project.domain.Project;
 import com.whatpl.domain.project.domain.ProjectLike;
-import com.whatpl.domain.project.repository.ProjectLikeRepository;
-import com.whatpl.domain.project.repository.ProjectRepository;
+import com.whatpl.domain.project.repository.like.ProjectLikeRepository;
+import com.whatpl.domain.project.repository.project.ProjectRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
