+++ src/main/java/com/whatpl/domain/project/service/ProjectParticipantService.java
@@ -6,9 +6,9 @@ import com.whatpl.global.exception.BizException;
 import com.whatpl.global.exception.ErrorCode;
 import com.whatpl.domain.project.domain.Project;
 import com.whatpl.domain.project.domain.ProjectParticipant;
-import com.whatpl.domain.project.repository.ApplyRepository;
-import com.whatpl.domain.project.repository.ProjectParticipantRepository;
-import com.whatpl.domain.project.repository.ProjectRepository;
+import com.whatpl.domain.project.repository.apply.ApplyRepository;
+import com.whatpl.domain.project.repository.participant.ProjectParticipantRepository;
+import com.whatpl.domain.project.repository.project.ProjectRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
