+++ src/main/java/com/whatpl/domain/project/service/ProjectReadService.java
@@ -7,8 +7,8 @@ import com.whatpl.domain.project.domain.Project;
 import com.whatpl.domain.project.dto.ProjectInfo;
 import com.whatpl.domain.project.dto.ProjectReadResponse;
 import com.whatpl.domain.project.dto.ProjectSearchCondition;
-import com.whatpl.domain.project.repository.ProjectLikeRepository;
-import com.whatpl.domain.project.repository.ProjectRepository;
+import com.whatpl.domain.project.repository.like.ProjectLikeRepository;
+import com.whatpl.domain.project.repository.project.ProjectRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Pageable;
