+++ src/main/java/com/whatpl/global/security/permission/manager/ApplyPermissionManager.java
@@ -6,8 +6,8 @@ import com.whatpl.global.security.domain.MemberPrincipal;
 import com.whatpl.domain.project.domain.Apply;
 import com.whatpl.domain.project.domain.Project;
 import com.whatpl.domain.project.model.ApplyType;
-import com.whatpl.domain.project.repository.ApplyRepository;
-import com.whatpl.domain.project.repository.ProjectRepository;
+import com.whatpl.domain.project.repository.apply.ApplyRepository;
+import com.whatpl.domain.project.repository.project.ProjectRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
