+++ src/main/java/com/whatpl/global/security/permission/manager/ProjectLikePermissionManager.java
@@ -4,7 +4,7 @@ import com.whatpl.global.exception.BizException;
 import com.whatpl.global.exception.ErrorCode;
 import com.whatpl.global.security.domain.MemberPrincipal;
 import com.whatpl.domain.project.domain.ProjectLike;
-import com.whatpl.domain.project.repository.ProjectLikeRepository;
+import com.whatpl.domain.project.repository.like.ProjectLikeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
