+++ src/test/java/com/whatpl/domain/project/service/ProjectParticipantServiceTest.java
@@ -9,9 +9,9 @@ import com.whatpl.domain.project.domain.ProjectParticipant;
 import com.whatpl.global.common.model.ApplyStatus;
 import com.whatpl.domain.project.model.ApplyFixture;
 import com.whatpl.domain.project.model.ProjectFixture;
-import com.whatpl.domain.project.repository.ApplyRepository;
-import com.whatpl.domain.project.repository.ProjectParticipantRepository;
-import com.whatpl.domain.project.repository.ProjectRepository;
+import com.whatpl.domain.project.repository.apply.ApplyRepository;
+import com.whatpl.domain.project.repository.participant.ProjectParticipantRepository;
+import com.whatpl.domain.project.repository.project.ProjectRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
